requirements.txt
alabaster==0.7.12
attrs==19.2.0
audioread==2.1.8
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
commonmark==0.9.1
cycler==0.10.0
docutils==0.15.2
fuzzywuzzy==0.17.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
Jinja2==2.10.3
jsonschema==3.1.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
more-itertools==7.2.0
numpy==1.17.3
packaging==19.2
pika==1.1.0
purepng==0.2.0
pyacoustid==1.1.7
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2019.3
recommonmark==0.6.0
requests==2.22.0
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.3.1
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
urllib3==1.25.6
zipp==0.6.0
